Remove the MF recoverer entirely.
Although MF is faster than CPCTPlus, it's also considerably more complex, and
I no longer think that complexity is worth it. We removed MF from the
documentation some while back so it's unlikely that anyone's using it.
