+++ nimbleparse/src/main.rs
@@ -75,11 +75,9 @@ fn main() {
     let quiet = matches.opt_present("q");
 
     let recoverykind = match matches.opt_str("r") {
-        None => RecoveryKind::MF,
+        None => RecoveryKind::CPCTPlus,
         Some(s) => match &*s.to_lowercase() {
             "cpctplus" => RecoveryKind::CPCTPlus,
-            "mf" => RecoveryKind::MF,
-            "panic" => RecoveryKind::Panic,
             "none" => RecoveryKind::None,
             _ => usage(prog, &format!("Unknown recoverer '{}'.", s)),
         },
